requirements.txt
oteapi-core>=0.0.6
requests==2.26.0
